+++ crates/hyperpuzzle_view/src/simulation/mod.rs
@@ -106,7 +106,7 @@ pub struct PuzzleSimulation {
     pub last_log_file: Option<PathBuf>,
 
     /// Timestamp signature, if any.
-    pub tsa_signature_v2: Option<String>,
+    pub tsa_signature_v3: Option<String>,
     /// Whether the solve has been saved to the autonamed file after completion.
     ///
     /// This field is only intended for use by downstream users of the crate; it
@@ -175,7 +175,7 @@ impl PuzzleSimulation {
 
             last_log_file: None,
 
-            tsa_signature_v2: None,
+            tsa_signature_v3: None,
             saved_to_autonamed_file: false,
             leaderboard_url: None,
 
@@ -1008,7 +1008,8 @@ impl PuzzleSimulation {
             scramble: self.scramble.clone(),
             log,
             tsa_signature_v1: None,
-            tsa_signature_v2: dbg!(self.tsa_signature_v2.clone()),
+            tsa_signature_v2: None,
+            tsa_signature_v3: self.tsa_signature_v3.clone(),
         }
     }
     /// Loads a log file from a string.
@@ -1021,7 +1022,8 @@ impl PuzzleSimulation {
             scramble,
             log,
             tsa_signature_v1: _,
-            tsa_signature_v2,
+            tsa_signature_v2: _,
+            tsa_signature_v3,
         } = solve;
 
         log::trace!("Loading file ...");
@@ -1130,7 +1132,7 @@ impl PuzzleSimulation {
         if ret.solved {
             ret.has_been_reloaded_since_first_solved = true;
         }
-        ret.tsa_signature_v2 = tsa_signature_v2.clone();
+        ret.tsa_signature_v3 = tsa_signature_v3.clone();
 
         ret.skip_twist_animations();